feat: attach built vscode extension and standalone js file to release
The release of the vscode extension happens when this workflow runs, which is normally on a merged pr from the changeset bot. Normally, this is what happens, but there still can be cases where it runs when a PR is created, if another changeset file gets into the sourcecode when this runs.